requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/requirements.txt requirements/requirements.in
#
boto3==1.9.86
boto==2.49.0              # via django-ses
botocore==1.12.86         # via boto3, s3transfer
certifi==2018.10.15       # via requests
chardet==3.0.4            # via requests
configparser==3.5.0
decorate-url==1.1
dj-database-url==0.5.0
django-appconf==1.0.2     # via django-axes
django-axes==4.4.0
django-bootstrap-form==3.4
django-cors-headers==2.4.0
django-dotenv==1.4.2
django-filter==2.0.0
django-getenv==1.3.2
django-ipware==2.1.0      # via django-axes
django-localflavor==1.6.2
django-oauth-toolkit==1.1.3
django-ses==0.8.7
django-settings-export==1.2.1
django-storages==1.7.1
django==2.1.2
djangorestframework-csv==2.1.0
djangorestframework==3.8.2
docutils==0.14            # via botocore
future==0.16.0            # via django-ses
idna==2.7                 # via requests
jmespath==0.9.3           # via boto3, botocore
jsonschema==2.6.0
newrelic==4.4.1.104
oauthlib==2.1.0
pillow==5.4.1
psycopg2==2.7.5
pyjwt==1.6.4
python-dateutil==2.7.5    # via botocore
python-openid==2.2.5
pytz==2018.5
requests-oauthlib==1.0.0
requests==2.20.0
s3transfer==0.1.13        # via boto3
six==1.11.0               # via djangorestframework-csv, six
sqlparse==0.2.4
unicodecsv==0.14.1        # via djangorestframework-csv
urllib3==1.23             # via requests
